Abstract logic from RSValue_Cmp into Rust function usable directly from Rust
