+++ src/redisearch_rs/c_entrypoint/rlookup_ffi/Cargo.toml
@@ -20,7 +20,7 @@ rlookup.workspace = true
 search_result.workspace = true
 sorting_vector.workspace = true
 value.workspace = true
-value_ffi = { path = "../value_ffi" }
+query_error.workspace = true
 workspace_hack.workspace = true
 
 [dev-dependencies]
